rename files to follow convertions
