+++ matchMultipleTemplate/main.py
@@ -1,6 +1,6 @@
 from src.logger import SingletonLogger
-from src.findTemplateImages import find_template_images
-from src.template_matching import templateMatching
+from src.findtemplateimages import find_template_images
+from src.templatematching import template_matching
 import argparse
 
 if __name__ == "__main__":
@@ -18,4 +18,4 @@ if __name__ == "__main__":
     directory = "et"
     logger = SingletonLogger(args.verbose, args.debug)
     find_template_images(args.directory, args.file)
-    templateMatching(args.directory, args.file)
+    template_matching(args.directory, args.file)
