+++ matchMultipleTemplate/src/geneticalgorithm/actualStiches/teste1.png
