+++ matchMultipleTemplate/src/geneticalgorithm/actualStiches/teste1_original.png
