+++ matchMultipleTemplate/src/geneticalgorithm/actualStiches/teste2.png
