+++ matchMultipleTemplate/src/geneticalgorithm/actualStiches/teste3.png
