+++ matchMultipleTemplate/src/logger.py
@@ -38,39 +38,10 @@ class SingletonLogger(object):
                 print(Fore.BLUE + f'{self.severity_info[severity]}: {newMessage}')
         elif severity == "VERBOSE":
             if self.__VERBOSE__:
-                # print()
                 print(Fore.MAGENTA + f'{self.severity_info[severity]}: {newMessage}')
-                # print(Style.RESET_ALL)
         elif severity == "WARNING":
             print(Fore.YELLOW + f'{self.severity_info[severity]}: {newMessage}')
         elif severity == "ERROR":
             print(Fore.RED + f'{self.severity_info[severity]}: {newMessage}')
         else:
             print(Fore.GREEN + f'{self.severity_info["INFO"]}: {newMessage}')
-
-
-# def logger.log(message, severity="INFO"):
-#     """Log a message to the console.
-#
-#     :param severity: The severity of the message.
-#     :param message: The message to log.
-#     """
-#     just_fix_windows_console()
-#
-#     now = datetime.datetime.now()
-#     newMessage = f"[{str(now)}] {message}"
-#
-#     if severity == 'DEBUG':
-#         if __DEBUG__:
-#             print(Fore.BLUE + f'{severity_info[severity]}: {newMessage}')
-#     elif severity == "VERBOSE":
-#         if __VERBOSE__:
-#             # print()
-#             print(Fore.MAGENTA + f'{severity_info[severity]}: {newMessage}')
-#             # print(Style.RESET_ALL)
-#     elif severity == "WARNING":
-#         print(Fore.YELLOW + f'{severity_info[severity]}: {newMessage}')
-#     elif severity == "ERROR":
-#         print(Fore.RED + f'{severity_info[severity]}: {newMessage}')
-#     else:
-#         print(Fore.GREEN + f'{severity_info["INFO"]}: {newMessage}')
