+++ matchMultipleTemplate/src/templatematching.py
@@ -1,15 +1,7 @@
 # -*- coding: utf-8 -*-
-"""How to use template matching with OpenCV? | Python
-
-Automatically generated by Colaboratory.
-
-Original file is located at
+"""
+    This was inspired by the following collab:
     https://colab.research.google.com/github/learn-ai-python/OpenCV-Tutorials/blob/main/Template_Matching/How_to_use_template_matching_with_OpenCV%3F_%7C_Python.ipynb
-
-# OpenCV template matching
-Hi! In this my new video we're going to learn how to use template mathing from OpenCV library.
-For this task we need two images - the first one with template and the second one with the target image.
-So, let's start!
 """
 
 import os
@@ -18,21 +10,14 @@ import cv2
 import numpy as np
 from matplotlib import pyplot as plt
 
-# sys.path.append('../geneticAlgorithm')
-
-# from geneticAlgorithm.genetic_algorithm import genetic_algorithm
-# from geneticAlgorithm.tsp import TSP
-# from logger import SingletonLogger
-
-
 __DELETE_FILES__ = False
 
-from .geneticAlgorithm.genetic_algorithm import genetic_algorithm
-from .geneticAlgorithm.tsp import TSP
+from .geneticalgorithm.geneticalgorithm import genetic_algorithm
+from .geneticalgorithm.tsp import TSP
 from .logger import SingletonLogger
 
 
-def removeCoordinatesClose(input_list, threshold=(10, 10)):
+def remove_coordinates_close(input_list, threshold=(10, 10)):
     combos = itertools.combinations(input_list, 2)
     points_to_remove = [point2
                         for point1, point2 in combos
@@ -42,7 +27,7 @@ def removeCoordinatesClose(input_list, threshold=(10, 10)):
     return points_to_keep
 
 
-def drawPath(problem, permutation, shortestHamPath=True):
+def draw_path(problem, permutation, shortestHamPath=True):
     coords = []
 
     splitPermutation = permutation.index(len(permutation) - 1)
@@ -58,16 +43,16 @@ def drawPath(problem, permutation, shortestHamPath=True):
     plt.plot(xs, ys, marker='o')
 
 
-def matchTemplate(fileName, dir_name, original_file_name):
+def match_template(fileName, dir_name, original_file_name):
     logger = SingletonLogger()
     logger.log("Matching template", "VERBOSE")
-    img = cv2.imread(dir_name+original_file_name)
+    img = cv2.imread(dir_name + original_file_name)
     # convert it from BGR to RGB
     imgRGB = cv2.cvtColor(img, cv2.COLOR_BGR2RGB)
     # and convert it from BGR to GRAY
     imgGray = cv2.cvtColor(img, cv2.COLOR_BGR2GRAY)
 
-    template = cv2.imread(dir_name+"/templates/"+fileName, 0)
+    template = cv2.imread(dir_name + "/templates/" + fileName, 0)
 
     # then we get the shape of the template
     w, h = template.shape[::-1]
@@ -83,7 +68,7 @@ def matchTemplate(fileName, dir_name, original_file_name):
 
     logger.log("DONE -- Matching template", "VERBOSE")
     # remove points too close to each other, likely the same image matched twice
-    newPoints = removeCoordinatesClose(list(zip(*loc[::-1])))
+    newPoints = remove_coordinates_close(list(zip(*loc[::-1])))
 
     tspFileName = "teste.tsp"
     f = open(tspFileName, "w")
@@ -112,24 +97,22 @@ def matchTemplate(fileName, dir_name, original_file_name):
 
     fig = plt.figure(figsize=(10, 10))
     plt.imshow(imgRGB, alpha=0.4)
-    drawPath(problem, path)
-    plt.savefig(dir_name+'/paths/' + fileName)
+    draw_path(problem, path)
+    plt.savefig(dir_name + '/paths/' + fileName)
     plt.close('all')
     os.remove(tspFileName)
 
 
-def templateMatching(directory, file_name):
+def template_matching(directory, file_name):
     entries = os.listdir(directory + "/templates/")
 
     if not os.path.isdir(directory + "/paths/"):
-        # if the demo_folder2 directory is
-        # not present then create it.
         os.makedirs(directory + "/paths/")
     files = os.listdir(directory + "/paths/")
+
     if __DELETE_FILES__:
         for f in files:
-            os.remove(directory + "/paths/"+f)
+            os.remove(directory + "/paths/" + f)
 
     for entry in entries:
-        matchTemplate(entry, directory, file_name)
-
+        match_template(entry, directory, file_name)
